pagination size: own component refactor
